correct ld not working

+++ projects/vm/incs/opcode.h
@@ -6,7 +6,7 @@
 /*   By: fbenneto <fbenneto@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/11 13:28:31 by glodi             #+#    #+#             */
-/*   Updated: 2019/03/15 15:05:38 by fbenneto         ###   ########.fr       */
+/*   Updated: 2019/03/21 16:21:20 by fbenneto         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,11 +20,12 @@
 #include "proc.h"
 #include "struct.h"
 #include "vm_utils.h"
+#include "flags.h"
 
 /**
  * utils read
  */
-uint16_t read_arg(t_process *process, uint8_t *mem, int type);
+uint32_t read_arg(t_process *process, uint8_t *mem, int type);
 uint8_t  read_octet_code(t_process *process, uint8_t *mem);
 void read_in_memory(uint8_t *memory, uint8_t *buffer, uint8_t len, size_t at);
